Added LikelyOutdatedMethodException
This is thrown from some methods that depend on string parsing.
Specifically, the methods that depend on very specific strings given from the OS may receive a different format than expected and would not know how to parse that string properly. This can and likely will be caused by the OS receiving an update that changes the expected string format.